# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/avatar": {
            "patch": {
                "description": "Change user avatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Change user avatar image",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/badge": {
            "get": {
                "description": "Get user's badges",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "Get user's badges",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Badge"
                        }
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login with a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "username and password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.loginUserType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.loginResponseType"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "Logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Logout failed"
                    }
                }
            }
        },
        "/pass": {
            "patch": {
                "description": "Change password given the old and new pass",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Change password of the user",
                "parameters": [
                    {
                        "description": "old and new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.changePasswordType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/player/{username}": {
            "get": {
                "description": "Get users' stats WITHOUT AUTH",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Get users' stats WITHOUT AUTH",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Stats"
                        }
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/player/{username}/avatar": {
            "get": {
                "description": "Return the player avatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Return the player avatar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username string",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "user with password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.customUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "user created",
                        "schema": {
                            "$ref": "#/definitions/types.User"
                        }
                    }
                }
            }
        },
...
        "/session": {
            "get": {
                "description": "Get auth session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Get current auth session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.User"
                        }
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "description": "Get users' stats",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get users' stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Stats"
                        }
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.changePasswordType": {
            "type": "object",
            "properties": {
                "new_pass": {
                    "type": "string"
                },
                "old_pass": {
                    "type": "string"
                }
            }
        },
...
        "handler.customUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "robot"
                },
                "firstname": {
                    "type": "string",
                    "example": "giorgio"
                },
                "lastname": {
                    "type": "string",
                    "example": "rossi"
                },
                "mail": {
                    "type": "string",
                    "example": "giorossi@mail.it"
                },
                "password": {
                    "type": "string",
                    "example": "1234"
                },
                "username": {
                    "type": "string",
                    "example": "gio"
                }
            }
        },
...
        "handler.loginResponseType": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "user": {
                    "$ref": "#/definitions/handler.loginResponseUser"
                }
            }
        },
        "handler.loginResponseUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "giorossi@mail.it"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "gio"
                }
            }
        },
        "handler.loginUserType": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "1234"
                },
                "username": {
                    "type": "string",
                    "example": "gio"
                }
            }
        },
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "types.Badge": {
            "type": "object",
            "properties": {
                "bot": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "elo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "pieces": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "played": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "time": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "win": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "types.LeaderboardUser": {
            "type": "object",
            "properties": {
                "elo": {
                    "type": "integer",
                    "example": 1000
                },
                "username": {
                    "type": "string",
                    "example": "gio"
                }
            }
        },
...
        "types.ReturnGame": {
            "type": "object",
            "properties": {
                "current_player": {
                    "type": "string",
                    "example": "p1"
                },
                "double_owner": {
                    "type": "string",
                    "example": "all"
                },
                "double_value": {
                    "type": "integer",
                    "example": 1
                },
                "elo1": {
                    "type": "integer",
                    "example": 1000
                },
                "elo2": {
                    "type": "integer",
                    "example": 1000
                },
                "end": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "game_type": {
                    "type": "string",
                    "example": "online"
                },
                "id": {
                    "type": "integer"
                },
                "p1checkers": {
                    "description": "arr[0] is bar",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "p2checkers": {
                    "description": "arr[0] is bar",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "player1": {
                    "description": "Username of the player",
                    "type": "string",
                    "example": "Giorgio"
                },
                "player2": {
                    "type": "string",
                    "example": "Mario"
                },
                "start": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "open"
                },
                "tournament": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "want_to_double": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "types.Stats": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "integer"
                },
                "elo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "games_played": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ReturnGame"
                    }
                },
                "leaderboard": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.LeaderboardUser"
                    }
                },
                "local": {
                    "type": "integer"
                },
                "lost": {
                    "type": "integer"
                },
                "online": {
                    "type": "integer"
                },
                "tournament": {
                    "type": "integer"
                },
                "win": {
                    "type": "integer"
                },
                "winrate": {
                    "type": "number"
                }
            }
        },
...
        "types.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": "robot"
                },
                "elo": {
                    "type": "integer",
                    "example": 1000
                },
                "firstname": {
                    "type": "string",
                    "example": "giorgio"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_bot": {
                    "type": "boolean",
                    "example": false
                },
                "lastname": {
                    "type": "string",
                    "example": "rossi"
                },
                "mail": {
                    "type": "string",
                    "example": "giorossi@mail.it"
                },
                "username": {
                    "type": "string",
                    "example": "gio"
                }
            }
        }
    }
}